Refactoring Routing, moving AiGenerateRoutes to a separated file. Introducing Social Content Profile. It is required to implement the scheduling of bumping content in the linkedin personal profile.
