+++ AI4Media/backend/AI4MediaServer/src/main/kotlin/com/catharsis/ai4media/ai4mediaserver/Application.kt
@@ -229,4 +229,5 @@ fun Application.module() {
     this.configureNewsRouting()
     this.configureScheduledContentRouting()
     this.configureReadingListRouting()
+    this.configureAIGenerationRouting()
 }
